Moved App.js to src folder
