+++ src/App.js
@@ -4,14 +4,15 @@ import axios from "axios";
 
 import styled from "styled-components";
 
-import "../App.css";
-import Home from "./Home.js";
-import OrderForm from "./OrderForm.js";
-import formSchema from "../validation/formSchema.js";
+import "./App.css";
+import Home from "./components/Home.js";
+import OrderForm from "./components/OrderForm.js";
+import formSchema from "./validation/formSchema.js";
 
 
 const App = () => {
 
+  // all values in the form are set to empty strings or to false (for checkboxes)
   const initialFormValues = {
     name: "",
     size: "",
@@ -24,6 +25,7 @@ const App = () => {
     special: "",
   }
 
+  // all errors are set to empty strings to start with
   const initialErrors = {
     name: "",
     size: "",
@@ -32,39 +34,40 @@ const App = () => {
   }
 
   // STATE
+    // the values of what is currently in the form of OrderForm.js
   const [formValues, setFormValues] = useState(initialFormValues);
+    // the errors occuring from validating the formValues, which are strings from formSchema.js
   const [errors, setErrors] =  useState(initialErrors);
+    // a total list of all the pizzas added to order
   const [orders, setOrders] = useState([]);
+    // disabled status of the submit button
   const [disabled, setDisabled] = useState(false);
 
+  // what happens when the "Submit" button is clicked
   const handleSubmit = event => {
+    postNewOrder(formValues);
+  }
+
+
+  // TODO - This adds a new order to the orders array, and ideally it will also post to an external API.
+  const postNewOrder = newOrder => {
     axios.post('https://reqres.in/api/orders', formValues)
       .then((res) => {
         setOrders([res.data, ...orders]);
       })
       .catch(err => console.error(err));
       setFormValues(initialFormValues);
+
   }
 
-  useEffect(() => {
-    axios.get("https://reqres.in/api/orders")
-      .then(res => {
-        console.log("Get results, ", res)
-      })
-      .catch(err => console.error(err));
-    })
-
-
+  // this updates the formValues with each change to the form, and also updates the errors if a formValue does not fit the requirements.
   const handleChange = (name, value) => {
     // validate here
     // console.log(`name: ${name}; value: ${value}`)
     setFormValues({...formValues, [name]: value})
   }
 
-  // useEffect(() => {
-  //   console.log(orders);
-  // }, [orders])
-
+  // sets the disabled status of the Submit button on the form
   useEffect(() => {
     formSchema.isValid()
       .then(valid => {
@@ -72,7 +75,13 @@ const App = () => {
       })
   }, [formValues])
 
+    // testing arry that prints a list of all the orders to the console
+  // useEffect(() => {
+  //   console.log(orders);
+  // }, [orders])
 
+
+  // JSX return
   return (
     <>
       <h1>Lambda Eats</h1>
