+++ src/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
-import App from "./components/App.js";
+import App from "./App.js";
 
 import { BrowserRouter as Router } from "react-router-dom";
